Allow trailing spaces in HTTP_X_FORWARDED_FOR (eg NetScaler MPX-15000)
We're using a NetScaler MPX-15000 (running NS9.3: Build 59.5.nc)
and it's adding trailing spaces into the X-Forwarded-For header.

This change allows trailing spaces to be ignored.

(It also prevents REMOTE_ADDR from being set to undef if the
HTTP_X_FORWARDED_FOR parsing fails. That seems like a good idea, but
isn't an important change for me.)